Use libvulkan for platform variant of deqp_amber
The SDK variant of deqp_amber doesn't need libvulkan, as the
headers are always visible through the NDK.  When the build
system is changed to also create a platform variant then
a dependency on libvulkan will be necessary to see the
headers.  Also reexport the headers to be visible to
libdeqp.

Bug: 149591340
Test: m checkbuild
Change-Id: I22cc9c2cea9e7aef7bcfc343c95e522b26ec6638