Update Android.bp
We now generate vk-wrappers.{inc,h} as part of the build, instead of
checking them in from an upstream build.

We no longer link with Vulkan, and instead define
-DAMBER_CTS_VULKAN_HEADER=1 (in deqp_and_deps_defaults) and use the
"Vulkan headers" from dEQP, as is done in the upstream Android dEQP
build.

All other defines are also moved to deqp_and_deps_defaults.

.cc files updated from the upstream Android dEQP build logs.

Bug: b/169021525
Change-Id: Id1e6fa091d5cf42e5de34db917effeba996e36eb